makefile: Fix linking error
The command line was passing all the object files
defined as dependency to the linker,
without removing duplicates.
Thus, the same name was being defined twice,
from the point of view of the linker ---
altough what was really hapenning is that the same object file
was being passed twice to the linker.